refactor : 유틸클래스 생성할 수 없도록 방어로직 추가
